third_party/WebKit/LayoutTests/platform/win/http/tests/usb/secure-context-expected.txt

This is a testharness.js-based test.
PASS navigator.usb is present in an insecure context 
Harness: the test ran to completion.
